Telco Customer Churn-ML Programming
You have to be careful about dividing your data correctly.

The data is given to us in a single file, we break this file ourselves and create the test and validation data, and thus we train this model by establishing a model from the data sets we have broken down.